+++ server/src/main/java/services/DeleteService.java
@@ -9,8 +9,13 @@ public class DeleteService {
     static MemUserDatabase userdatabase = MemDatabaseRegistry.getUserDb();
     static MemAuthDatabase authdatabase = MemDatabaseRegistry.getAuthDb();
     static MemGameDatabase gamedatabase = MemDatabaseRegistry.getGameDb();
+    private static boolean isMemoryImplemntation = DatabaseConfigLoader.memoryimplementation();
 
     public static boolean deleteAll() {
-        return userdatabase.deleteall() && authdatabase.deleteall() && gamedatabase.deleteall();
+        if(isMemoryImplemntation) {
+            return userdatabase.deleteall() && authdatabase.deleteall() && gamedatabase.deleteall();
+        }
+
+        return false;
     }
 }
