+++ server/src/main/java/services/GameService.java
@@ -10,10 +10,13 @@ import java.util.ArrayList;
 public class GameService {
     MemGameDatabase gameDatabase = MemDatabaseRegistry.getGameDb();
     private int gameID = 0;
-
+    private boolean isMemoryImplemntation = DatabaseConfigLoader.memoryimplementation();
 
     public boolean checkGameID(int gameID) {
-        return (gameDatabase.inDatabaseID(gameID) != null);
+        if(isMemoryImplemntation) {
+            return (gameDatabase.inDatabaseID(gameID) != null);
+        }
+        return false;
     }
 
 
@@ -23,48 +26,56 @@ public class GameService {
         }
         gameID++;
         int gameid = gameID;
-        GameData gameData = new GameData(gameid, null, null, gameName, new ChessGame());
-        if (gameDatabase.addToDatabase(gameData)) {
-            return gameData;
+        if(isMemoryImplemntation) {
+            GameData gameData = new GameData(gameid, null, null, gameName, new ChessGame());
+            if (gameDatabase.addToDatabase(gameData)) {
+                return gameData;
+            }
+            return null;
         }
         return null;
-
     }
 
     public ArrayList<GameData> getGames() {
+        if(isMemoryImplemntation){
         return gameDatabase.getDatabase();
+    }
+        return null;
     }
 
     public boolean assignColor(String username, ChessGame.TeamColor joinColor, int gameID) {
-        GameData data = gameDatabase.inDatabaseID(gameID);
-        if (joinColor == ChessGame.TeamColor.WHITE) {
-            if (data.whiteUsername() == null) {
-                GameData updatedData = new GameData(
-                        data.gameID(),
-                        username,               // new whiteUsername
-                        data.blackUsername(),
-                        data.gameName(),
-                        data.game()
-                );
-                gameDatabase.removeFromDatabase(data);
-                gameDatabase.addToDatabase(updatedData);
-                return true;
+        if(isMemoryImplemntation){
+            GameData data = gameDatabase.inDatabaseID(gameID);
+            if (joinColor == ChessGame.TeamColor.WHITE) {
+                if (data.whiteUsername() == null) {
+                    GameData updatedData = new GameData(
+                            data.gameID(),
+                            username,               // new whiteUsername
+                            data.blackUsername(),
+                            data.gameName(),
+                            data.game()
+                    );
+                    gameDatabase.removeFromDatabase(data);
+                    gameDatabase.addToDatabase(updatedData);
+                    return true;
+                }
             }
-        }
-        if (joinColor == ChessGame.TeamColor.BLACK) {
-            if (data.blackUsername() == null) {
-                GameData updatedData = new GameData(
-                        data.gameID(),
-                        data.whiteUsername(),               // new whiteUsername
-                        username,
-                        data.gameName(),
-                        data.game()
-                );
-                gameDatabase.removeFromDatabase(data);
-                gameDatabase.addToDatabase(updatedData);
-                return true;
+            if (joinColor == ChessGame.TeamColor.BLACK) {
+                if (data.blackUsername() == null) {
+                    GameData updatedData = new GameData(
+                            data.gameID(),
+                            data.whiteUsername(),               // new whiteUsername
+                            username,
+                            data.gameName(),
+                            data.game()
+                    );
+                    gameDatabase.removeFromDatabase(data);
+                    gameDatabase.addToDatabase(updatedData);
+                    return true;
 
+                }
             }
+            return false;
         }
         return false;
     }
